Fix API - 로드할 API 정리
